Wycheproof EDDSA: re-work
There are two macros now. The first one is a function like macro for
generating the equivalent of the JSON data as static Rust data during
compilation. This can be used e.g. for a no_std on-target test without
the need for further runtime parsing. (See
qemu-tests/src/bin/wycheproof_on_target/main.rs and `make test' in
qemu-tests.)

The other one is a derive macro to be applied to a test function that
handles one EDDSA test. A call marked [#test] to this function will be
generated for each test case. This way, proper Rust tests can be
generated. (See tests/wycheproof_eddsa.rs and `make test' in the main
salty directory.)